ai-service/requirements.txt
fastapi>=0.115,<0.117
uvicorn[standard]>=0.34,<0.36
pydantic>=2.10,<3
python-dotenv>=1.0,<2
langgraph>=0.2,<1
langchain-anthropic>=0.3,<1
langchain-openai>=0.3,<1
httpx>=0.27,<1
